Added scrollable chat views
