+++ landline/landline/Base.lproj/Main.storyboard
@@ -111,8 +111,23 @@
                     <view key="view" contentMode="scaleToFill" id="kHk-Cu-LNj">
                         <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <scrollView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Y2c-vP-JcQ">
+                                <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
+                            </scrollView>
+                        </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <constraints>
+                            <constraint firstItem="Y2c-vP-JcQ" firstAttribute="leading" secondItem="kHk-Cu-LNj" secondAttribute="leading" id="6sK-wr-AR0"/>
+                            <constraint firstAttribute="trailing" secondItem="Y2c-vP-JcQ" secondAttribute="trailing" id="bGX-Yd-ggz"/>
+                            <constraint firstItem="Y2c-vP-JcQ" firstAttribute="top" secondItem="kHk-Cu-LNj" secondAttribute="top" id="lwK-EJ-QLb"/>
+                            <constraint firstAttribute="bottom" secondItem="Y2c-vP-JcQ" secondAttribute="bottom" id="oWq-0j-3gj"/>
+                        </constraints>
                     </view>
+                    <simulatedNavigationBarMetrics key="simulatedTopBarMetrics" prompted="NO"/>
+                    <connections>
+                        <outlet property="channelScrollView" destination="Y2c-vP-JcQ" id="6Qr-fs-3C2"/>
+                    </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="hrz-Vw-3JK" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
@@ -326,6 +341,24 @@
                 <placeholder placeholderIdentifier="IBFirstResponder" id="BUI-XC-Bfx" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="1008" y="322"/>
+        </scene>
+        <!--Channel View Controller-->
+        <scene sceneID="LJQ-fV-sbf">
+            <objects>
+                <viewController storyboardIdentifier="ChannelViewController" id="9K9-bC-iy2" customClass="ChannelViewController" customModule="landline" customModuleProvider="target" sceneMemberID="viewController">
+                    <layoutGuides>
+                        <viewControllerLayoutGuide type="top" id="ZRn-z2-ptl"/>
+                        <viewControllerLayoutGuide type="bottom" id="4fL-Pk-Uxe"/>
+                    </layoutGuides>
+                    <view key="view" contentMode="scaleToFill" id="FfY-dY-IY3">
+                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="8RV-bC-SPi" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="1008" y="1044"/>
         </scene>
         <!--Settings-->
         <scene sceneID="JeA-vb-gIe">
